+++ backend/package.json
@@ -2,11 +2,13 @@
   "name": "sakura-restaurant-backend",
   "version": "1.0.0",
   "description": "Backend API for Sakura Japanese Restaurant QR Menu System",
-  "main": "server.js",
+  "main": "dist/server.js",
+  "type": "module",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js",
-    "seed": "node scripts/seed-database.js",
+    "build": "tsc",
+    "start": "node dist/server.js",
+    "dev": "tsx watch server.ts",
+    "seed": "tsx scripts/seed-database.js",
     "test": "jest"
   },
   "dependencies": {
@@ -25,11 +27,20 @@
     "moment": "^2.29.4"
   },
   "devDependencies": {
+    "@types/express": "^4.17.21",
+    "@types/cors": "^2.8.17",
+    "@types/node": "^20.10.0",
+    "@types/bcryptjs": "^2.4.6",
+    "@types/jsonwebtoken": "^9.0.5",
+    "@types/multer": "^1.4.11",
+    "@types/nodemailer": "^6.4.14",
+    "typescript": "^5.3.0",
+    "tsx": "^4.6.0",
     "nodemon": "^3.0.2",
     "jest": "^29.7.0",
     "supertest": "^6.3.3"
   },
-  "keywords": ["restaurant", "qr-menu", "api", "mongodb", "express"],
+  "keywords": ["restaurant", "qr-menu", "api", "mongodb", "express", "typescript"],
   "author": "Your Name",
   "license": "MIT"
 }
