--- frontend/app/api/orders/%5Bid%5D/route.ts
@@ -1,3 +1,0 @@
-import dbConnect from "@/lib/mongodb"
-import Order from "@/models1/Order"
-import Table from "@/models1/Table"
